+++ apps/dotcom/src/components/Timer.ts
@@ -4,7 +4,7 @@ export const showTimer = atom('timer', false)
 
 export function setTimer(editor: Editor) {
 	let meta = editor.getDocumentSettings().meta as any
-	if (!meta.timer || !meta.timer.initialTime) {
+	if (!meta.timer || meta.timer.initialTime === undefined) {
 		meta = {
 			...meta,
 			timer: {
